backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary
google-generativeai
python-dotenv
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
werkzeug
gunicorn